+++ source/Playnite/App/PlayniteApplication.cs
@@ -342,6 +342,13 @@ namespace Playnite
             SDK.Data.Serialization.Init(new DataSerializer());
             Startup();
             logger.Info($"Application {CurrentVersion} started");
+            foreach (var fail in Extensions.FailedExtensions)
+            {
+                Api.Notifications.Add(new NotificationMessage(
+                    fail.DirectoryPath,
+                    ResourceProvider.GetString(LOC.SpecificExtensionLoadError).Format(fail.Name),
+                    NotificationType.Error));
+            }
         }
 
         private void PipeService_CommandExecuted(object sender, CommandExecutedEventArgs args)
@@ -484,8 +491,8 @@ namespace Playnite
                 catch (Exception exc) when (!PlayniteEnvironment.ThrowAllErrors)
                 {
                     MessageBox.Show(
-                        ResourceProvider.GetString("LOCStartGenericError"),
-                        ResourceProvider.GetString("LOCStartupError"));
+                        "Playnite failed to start. Please close all other instances and try again.",
+                        "Startup Error");
                     logger.Error(exc, "Can't process communication with other instances.");
                 }