+++ source/Tools/Playnite.Toolbox/Extensions.cs
@@ -23,7 +23,7 @@ namespace Playnite.Toolbox
 
         public static string GenerateScriptExtension(ScriptLanguage language, string name, string directory)
         {
-            var extDirName = Common.Paths.GetSafeFilename(name).Replace(" ", string.Empty);
+            var extDirName = Common.Paths.GetSafePathName(name).Replace(" ", string.Empty);
             var outDir = Path.Combine(directory, extDirName);
             if (Directory.Exists(outDir))
             {
@@ -37,7 +37,7 @@ namespace Playnite.Toolbox
 
         public static string GeneratePluginExtension(ExtensionType type, string name, string directory)
         {
-            var normalizedName = Common.Paths.GetSafeFilename(name).Replace(" ", string.Empty);
+            var normalizedName = Common.Paths.GetSafePathName(name).Replace(" ", string.Empty);
             var outDir = Path.Combine(directory, normalizedName);
             if (Directory.Exists(outDir))
             {
@@ -108,7 +108,7 @@ namespace Playnite.Toolbox
 
             extInfo.VerifyManifest();
 
-            var packedPath = Path.Combine(targetPath, $"{Common.Paths.GetSafeFilename(extInfo.Name).Replace(' ', '_')}_{extInfo.Version.ToString().Replace(".", "_")}{PlaynitePaths.PackedExtensionFileExtention}");
+            var packedPath = Path.Combine(targetPath, $"{Common.Paths.GetSafePathName(extInfo.Name).Replace(' ', '_')}_{extInfo.Version.ToString().Replace(".", "_")}{PlaynitePaths.PackedExtensionFileExtention}");
             FileSystem.PrepareSaveFile(packedPath);
             var ignoreFiles = File.ReadAllLines(Paths.ExtFileIgnoreListPath);